[vnc] add VNC file transfer support with UVNC server (WAB-12931, WAB-14069, WAB-14070, WAB-14071, WAB-14077, WAB-15114)

+++ src/utils/parse.hpp
@@ -154,7 +154,7 @@ public:
         this->p += v.size();
     }
 
-    void in_skip_bytes(unsigned int n) noexcept {
+    void in_skip_bytes(size_t n) noexcept {
         this->p+=n;
     }
